# [SPARK-25038][SQL] Get block location in parallel