More details please refer to
https://issues.apache.org/jira/browse/SPARK-25038

When Spark SQL read large amount of data, it take a long time (more than 10 minutes) to generate physical Plan and then ActiveJob because sequentially get block location is very time-consuming. 

To improve this, ParArray can be used to get block location in parallel

 

Example:

There is a table which is partitioned by date and hour. There are more than 13 TB data each hour and 185 TB per day. When we just issue a very simple SQL, it take a long time to generate ActiveJob

 

The SQL statement is

select count(device_id) from test_tbl where date=20180731 and hour='21';
 

Before optimization, it takes 2 minutes and 9 seconds to generate the Job

 

The SQL is issued at 2018-08-07 09:07:41



However, the job is submitted at 2018-08-07 09:09:53, which is 2minutes and 9 seconds later than the SQL issue time



 

After the optimization, it takes only 4 seconds to generate the Job

The SQL is issued at 2018-08-07 09:20:15



 

And the job is submitted at 2018-08-07 09:20:19, which is 4 seconds later than the SQL issue time