@wangyum 2019-08-18
Hi @habren I think this performance issue has been fixed by [SPARK-27801](https://issues.apache.org/jira/browse/SPARK-27801). Could you close this?